Generalize and document itemized committee queries.
Apply the `UNION ALL` optimization for performant queries on itemized
resources by multiple committees to the base `ItemizedResource` class.
This patch improves performance for these queries against Schedule B
transactions.
